Fix compilation errors in Ubuntu.
Missing headers <algorithm>  and missing macro SIZE_MAX

+++ src/externDefinitions.h
@@ -59,7 +59,7 @@ namespace StochHMM{
     class ExDefSequence{
     public:
         //Constructor
-        ExDefSequence(int);
+        ExDefSequence(size_t);
         
         
         //Copy Constructor
